// src/app/model/douban.ts
export interface Douban {
  doubanId: number;
  ratingAverage: number;
  ratingNumRaters: number;
  author: string;
  altTitle: string;
  image: string;
  title: string;
  summary: string;
  tags: string;
  website: string;
  language: string;
  pubdate: string;
  country: string;
  writer: string;
  director: string;
  cast: string;
  movieDuration: string;
  movieType: string;
  year: string;
}