// src/app/model/imdb.ts
export interface Imdb {
  imdbId: number;
  title: string;
  year: string;
  rated: string;
  released: string;
  runtime: string;
  genre: string;
  director: string;
  writer: string;
  actors: string;
  plot: string;
  language: string;
  country: string;
  poster: string;
  ratingImd: number;
  ratingRt: number;
  ratingMetacritic: number;
  imdbRating: number;
  imdbVotes: number;
  type: string;
  production: string;
}